Add fetching stream info
This commit adds the functionality to fetch additional info from the
twitch "streams" API. This is then used to display streamers streaming
non-infosec stuff (filtered by the game name) as offline and also
display the listed language.

The updateStreamStatus() function was also completley rewritten to make
the additional displayed info easier to implement.