Add fetching stream info (#1)
* Add fetching stream info

This commit adds the functionality to fetch additional info from the
twitch "streams" API. This is then used to display streamers streaming
non-infosec stuff (filtered by the game name) as offline and also
display the listed language.

The updateStreamStatus() function was also completley rewritten to make
the additional displayed info easier to implement.

* Apply suggestions from code review

Co-authored-by: goproslowyo <68455785+goproslowyo@users.noreply.github.com>